+++ packages/react-components/src/index.ts
@@ -2,6 +2,7 @@ export * from './useConnect';
 export * from './useConnection';
 export * from './useContractSelector';
 export * from './useGrpcClient';
+export * from './useModuleSchemaRpc';
 export * from './useWalletConnectorSelector';
 export * from './WithWalletConnector';
 export * from '@concordium/wallet-connectors';