+++ packages/react-components/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+-   Hook `useModuleSchemaRpc` for fetching the schema of a smart contract module from the chain.
+
 ## [0.5.1] - 2024-03-22
 
 -   Dependency on `@concordium/wallet-connectors` bumped to v0.5.1+.